Enable inheriting from previously singleton filter and calc constant reduction rules.
Enables dynamic schema planning to override behavior in the case of a false filter
where a Values operator cannot be provided with a schema at planning time,
schema must propagate during execution.

To make the diff for the function change clean this is currently aligned poorly, will
add another patch to fix the formatting.

Formatting fixes to make the check style in the build pass.

Change rule name

Conflicts:
	core/src/main/java/org/apache/calcite/rel/rules/ReduceExpressionsRule.java